requirements.txt
# Core dependencies
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
delta-spark>=3.0.0
pyspark>=3.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
alpha-vantage>=2.3.0
fredapi>=0.5.0

# ML and AI
scikit-learn>=1.3.0
mlflow>=2.8.0
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
